pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-alpaca-trading-workflow"
version = "0.1.0"
description = "A modular Alpaca-based agentic trading workflow for market analysis, retrieval, decisioning, and backtesting."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alpaca-py",
  "matplotlib",
  "numpy",
  "pandas",
  "python-dotenv",
  "scikit-learn",
]

[project.optional-dependencies]
dev = ["pytest"]

[project.scripts]
agentic-trading = "agentic_trading.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

